Added solution - LeetHub

+++ Reverse an Array - GFG/reverse-an-array.cpp
@@ -3,20 +3,21 @@ using namespace std;
 
 int main() {
 	//code
-	long long t;
+	int t;
 	cin>>t;
 	while(t--){
-	    long long n;
+	    int n;
 	    cin>>n;
-	    long long arr[n];
-	    for(long long i=0;i<n;i++){
+	    int arr[n];
+	    for(int i=0;i<n;i++){
 	        cin>>arr[i];
 	    }
 	    
-	    for(long long i=n-1;i>=0;i--){
+	    for(int i=n-1;i>=0;i--){
 	        cout<<arr[i]<<" ";
 	    }
 	    cout<<endl;
+	    
 	}
 	return 0;
 }
